refactor(storage/tests): extract Shared-action signing helper into common (#2233 review pass 1)
Addresses the bot reviewer's DRY suggestion on PR #2265. Two near-identical
~60-line helpers existed:
- `build_signed_shared_action` in `tests/p3_dag_causal.rs`
- `build_signed` in `tests/p5_partition_scenarios.rs` (different name,
  identical body)

Both — and the `pubkey_of` helper they shared — now live in
`tests/common.rs` alongside the existing User-storage equivalents
(`create_signed_user_add_action` etc.). Callsites in p3 and p5 now
import via `use crate::tests::common::{build_signed_shared_action,
pubkey_of};`. p5's local callsites renamed `build_signed(` →
`build_signed_shared_action(` (17 sites) for consistency.

Also drops the now-unused imports both modules picked up to support
the local helpers (`Action`, `BTreeSet`, `PublicKey`, `SignatureData`,
`StorageType`, `Signer` — kept only those still referenced by test
bodies).

Tests: 429/429 storage lib tests still pass. Clippy on storage is clean.
